requirements.txt
boto3==1.26.28
click==8.1.3
pandas==1.5.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
sqlalchemy==1.4.45
